+++ apps/server/src/shared/domain/domainobject/board/card.do.ts
@@ -1,7 +1,7 @@
 import { BoardComposite, BoardCompositeProps } from './board-composite.do';
 import { FileElement } from './file-element.do';
 import { RichTextElement } from './rich-text-element.do';
-import { SubmissionContainerElement } from './task-element.do';
+import { SubmissionContainerElement } from './submission-container-element.do';
 import type { AnyBoardDo, BoardCompositeVisitor, BoardCompositeVisitorAsync } from './types';
 
 export class Card extends BoardComposite<CardProps> {
