+++ apps/server/src/shared/domain/domainobject/board/index.ts
@@ -5,5 +5,5 @@ export * from './column.do';
 export * from './content-element.factory';
 export * from './file-element.do';
 export * from './rich-text-element.do';
-export * from './task-element.do';
+export * from './submission-container-element.do';
 export * from './types';
