+++ apps/server/src/shared/domain/domainobject/board/types/board-composite-visitor.ts
@@ -3,7 +3,7 @@ import type { ColumnBoard } from '../column-board.do';
 import type { Column } from '../column.do';
 import type { FileElement } from '../file-element.do';
 import { RichTextElement } from '../rich-text-element.do';
-import { SubmissionContainerElement } from '../task-element.do';
+import { SubmissionContainerElement } from '../submission-container-element.do';
 
 export interface BoardCompositeVisitor {
 	visitColumnBoard(columnBoard: ColumnBoard): void;
